[Waste] Create shared redirect function.

+++ perllib/FixMyStreet/App/Controller/Waste.pm
@@ -822,13 +822,11 @@ sub enquiry : Chained('property') : Args(0) {
     my $category = $c->get_param('category');
     my $service = $c->get_param('service_id');
     if (!$category || !$service || !$c->stash->{services}{$service}) {
-        $c->res->redirect('/waste/' . $c->stash->{property}{id});
-        $c->detach;
+        $c->detach('property_redirect');
     }
     my ($contact) = grep { $_->category eq $category } @{$c->stash->{contacts}};
     if (!$contact) {
-        $c->res->redirect('/waste/' . $c->stash->{property}{id});
-        $c->detach;
+        $c->detach('property_redirect');
     }
 
     my $field_list = [];
@@ -903,15 +901,8 @@ sub check_if_staff_can_pay : Private {
 sub garden : Chained('property') : Args(0) {
     my ($self, $c) = @_;
 
-    if ($c->stash->{waste_features}->{garden_disabled}) {
-        $c->res->redirect('/waste/' . $c->stash->{property}{id});
-        $c->detach;
-    }
-
-    if ( $c->stash->{services}{$c->cobrand->garden_waste_service_id} ) {
-        $c->res->redirect('/waste/' . $c->stash->{property}{id});
-        $c->detach;
-    }
+    $c->detach('property_redirect') if $c->stash->{waste_features}->{garden_disabled};
+    $c->detach('property_redirect') if $c->stash->{services}{$c->cobrand->garden_waste_service_id};
 
     my $service = $c->cobrand->garden_waste_service_id;
     $c->stash->{garden_form_data} = {
@@ -926,14 +917,8 @@ sub garden : Chained('property') : Args(0) {
 sub garden_modify : Chained('property') : Args(0) {
     my ($self, $c) = @_;
 
-    if ($c->stash->{waste_features}->{garden_disabled}) {
-        $c->res->redirect('/waste/' . $c->stash->{property}{id});
-        $c->detach;
-    }
-
-    unless ( $c->user_exists ) {
-        $c->detach( '/auth/redirect' );
-    }
+    $c->detach('property_redirect') if $c->stash->{waste_features}->{garden_disabled};
+    $c->detach( '/auth/redirect' ) unless $c->user_exists;
 
     $c->forward('get_original_sub');
 
@@ -948,8 +933,7 @@ sub garden_modify : Chained('property') : Args(0) {
     my $max_bins = $c->stash->{quantity_max}->{$service};
     $service = $c->stash->{services}{$service};
     if (!$service || $service->{garden_due}) {
-        $c->res->redirect('/waste/' . $c->stash->{property}{id});
-        $c->detach;
+        $c->detach('property_redirect');
     }
 
     my $payment_method = 'credit_card';
@@ -981,19 +965,9 @@ sub garden_modify : Chained('property') : Args(0) {
 sub garden_cancel : Chained('property') : Args(0) {
     my ($self, $c) = @_;
 
-    if ($c->stash->{waste_features}->{garden_disabled}) {
-        $c->res->redirect('/waste/' . $c->stash->{property}{id});
-        $c->detach;
-    }
-
-    if ( !$c->stash->{services}{$c->cobrand->garden_waste_service_id} ) {
-        $c->res->redirect('/waste/' . $c->stash->{property}{id});
-        $c->detach;
-    }
-
-    unless ( $c->user_exists ) {
-        $c->detach( '/auth/redirect' );
-    }
+    $c->detach('property_redirect') if $c->stash->{waste_features}->{garden_disabled};
+    $c->detach('property_redirect') if !$c->stash->{services}{$c->cobrand->garden_waste_service_id};
+    $c->detach( '/auth/redirect' ) unless $c->user_exists;
 
     $c->forward('get_original_sub');
 
@@ -1008,14 +982,8 @@ sub garden_cancel : Chained('property') : Args(0) {
 sub garden_renew : Chained('property') : Args(0) {
     my ($self, $c) = @_;
 
-    if ($c->stash->{waste_features}->{garden_disabled}) {
-        $c->res->redirect('/waste/' . $c->stash->{property}{id});
-        $c->detach;
-    }
-
-    unless ( $c->user_exists ) {
-        $c->detach( '/auth/redirect' );
-    }
+    $c->detach('property_redirect') if $c->stash->{waste_features}->{garden_disabled};
+    $c->detach( '/auth/redirect' ) unless $c->user_exists;
 
     $c->forward('get_original_sub');
 
@@ -1556,6 +1524,12 @@ sub uprn_redirect : Path('/property') : Args(1) {
     $c->detach( '/page_error_404_not_found', [] ) unless $result;
     $c->res->redirect('/waste/' . $result->{Id});
 }
+
+sub property_redirect : Private {
+    my ($self, $c) = @_;
+    $c->res->redirect('/waste/' . $c->stash->{property}{id});
+}
+
 sub label_for_field {
     my ($form, $field, $key) = @_;
     foreach ($form->field($field)->options) {
